src/main/java/agh/cs/gameOfLife/Vector2d.java
package agh.cs.gameOfLife;

public class Vector2d {
    private final int x;
    private final int y;

    public Vector2d(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public int getX() {
        return x;
    }

    public int getY() {
        return y;
    }

    public String toString() {
        return "(" + this.x+ "," + this.y+ ")";
    }

    public boolean precedes(Vector2d other) {
        if (x <= other.x && x <= other.y && y <= other.x && y <= other.y) {
            return true;
        }
        else
            return false;
    }

    public boolean follows(Vector2d other) {
        if (x >= other.x && x >= other.y && y >= other.x && y >= other.y) {
            return true;
        }
        else
            return false;
    }

    public Vector2d upperRight(Vector2d other) {
        return new Vector2d(Math.max(this.x, other.x), Math.max(this.y, other.y));
    }

    public Vector2d lowerLeft(Vector2d other) {
        return new Vector2d(Math.min(x, other.x), Math.min(y, other.y));
    }

    public Vector2d add(Vector2d other) {
        return new Vector2d(this.x + other.x, this.y + other.y);
    }

    public Vector2d subtract(Vector2d other) {
        return new Vector2d(x - other.x, y - other.y);
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        Vector2d vector2d = (Vector2d) o;

        if (x != vector2d.x) return false;
        return y == vector2d.y;
    }

    @Override
    public int hashCode() {
        int hash = 13;
        hash += this.x * 31;
        hash += this.y * 17;
        return hash;
    }

    public Vector2d opposite() {
        return new Vector2d(-x, -y);
    }
}
